requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.6
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
backcall==0.2.0
billiard==3.6.4.0
black==21.11b1
bleach==4.1.0
ccxt==1.62.56
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
drf-yasg==1.20.0
emoji==1.6.1
entrypoints==0.3
frozenlist==1.2.0
idna==3.3
inflection==0.5.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
ntplib==0.4.0
numpy==1.21.5
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
pycares==4.1.2
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
redis==4.0.2
regex==2021.11.10
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.2
testpath==0.5.0
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
widgetsnbextension==3.5.2
wrapt==1.13.3
yarl==1.7.2
